avoid floats in MAF int cols

+++ pcgr/maf.py
@@ -7,12 +7,12 @@ import pandas as pd
 
 from pcgr.utils import check_file_exists, remove_file
 
-def update_maf_allelic_support(maf_tmp_fname: str, 
-                               maf_fname: str, 
-                               allelic_support_tags: dict,
-                               logger = None,
-                               update_allelic_support = False
-                              ):
+def update_maf(maf_tmp_fname: str, 
+               maf_fname: str, 
+               allelic_support_tags: dict,
+               logger = None,
+               update_allelic_support = False,
+               debug = False):
 
     """
     Update MAF file from vcf2maf.pl with allelic support data (t_depth, t_ref_count, t_alt_count etc).
@@ -35,7 +35,7 @@ def update_maf_allelic_support(maf_tmp_fname: str,
         header_line = f.readline().strip('\n')
     f.close()
         
-    raw_maf_data = pd.read_csv(maf_tmp_fname, sep="\t", header=1, na_values=['.'], low_memory=False)
+    raw_maf_data = pd.read_csv(maf_tmp_fname, sep="\t", header=1, dtype='string',na_values=['.'], low_memory=False)
     if update_allelic_support is False:
         # write to file
         b = 1
@@ -46,50 +46,47 @@ def update_maf_allelic_support(maf_tmp_fname: str,
             if allelic_support_tags['tumor_dp_tag'] != "_NA_":                               
                 if {allelic_support_tags['tumor_dp_tag']}.issubset(raw_maf_data.columns):
                     if raw_maf_data[raw_maf_data[allelic_support_tags['tumor_dp_tag']].isna() == True].empty is True:
-                        raw_maf_data = raw_maf_data.astype({allelic_support_tags['tumor_dp_tag']:'int'})                                        
                         raw_maf_data.loc[:,"t_depth"] = raw_maf_data.loc[:,allelic_support_tags['tumor_dp_tag']]
                         
                         if 'tumor_af_tag' in allelic_support_tags: 
                             if allelic_support_tags['tumor_af_tag'] != "_NA_":    
-                                if {allelic_support_tags['tumor_af_tag']}.issubset(raw_maf_data.columns):
-                                    
+                                if {allelic_support_tags['tumor_af_tag']}.issubset(raw_maf_data.columns):                                   
                                     if raw_maf_data[raw_maf_data[allelic_support_tags['tumor_af_tag']].isna() == True].empty is True:
+                                        raw_maf_data['t_alt_count'] = None
                                         raw_maf_data.loc[:,"t_alt_count"] = \
-                                            raw_maf_data.loc[:,allelic_support_tags['tumor_af_tag']] * \
-                                            raw_maf_data.loc[:,"t_depth"]
+                                            raw_maf_data.loc[:,allelic_support_tags['tumor_af_tag']].astype(float) * raw_maf_data.loc[:,"t_depth"].astype(int)
                                     
-                                        raw_maf_data.loc[:,"t_alt_count"] = raw_maf_data.loc[:,"t_alt_count"].round(0).astype(int)
-                                        
+                                        raw_maf_data.loc[:,"t_alt_count"] = round(raw_maf_data.loc[:,"t_alt_count"].astype(float),0).astype(int)
+                                        raw_maf_data['t_ref_count'] = None
                                         raw_maf_data.loc[:,"t_ref_count"] = \
-                                            raw_maf_data.loc[:,"t_depth"] - raw_maf_data.loc[:,"t_alt_count"]
+                                            raw_maf_data.loc[:,"t_depth"].astype(int) - raw_maf_data.loc[:,"t_alt_count"]
         
         if 'control_dp_tag' in allelic_support_tags:
             if allelic_support_tags['control_dp_tag'] != "_NA_":                    
                 if {allelic_support_tags['control_dp_tag']}.issubset(raw_maf_data.columns):
                     if raw_maf_data[raw_maf_data[allelic_support_tags['control_dp_tag']].isna() == True].empty is True:
-                        raw_maf_data = raw_maf_data.astype({allelic_support_tags['control_dp_tag']:'int'})                    
                         raw_maf_data.loc[:,"n_depth"] = raw_maf_data.loc[:,allelic_support_tags['control_dp_tag']]
                         
                         if 'control_af_tag' in allelic_support_tags: 
                             if allelic_support_tags['control_af_tag'] != "_NA_":    
                                 if {allelic_support_tags['control_af_tag']}.issubset(raw_maf_data.columns):
-                                    
                                     if raw_maf_data[raw_maf_data[allelic_support_tags['control_af_tag']].isna() == True].empty is True:
+                                        raw_maf_data['n_alt_count'] = None
                                         raw_maf_data.loc[:,"n_alt_count"] = \
-                                            raw_maf_data.loc[:,allelic_support_tags['control_af_tag']] * \
-                                            raw_maf_data.loc[:,"n_depth"]
-                                        
-                                        raw_maf_data.loc[:,"n_alt_count"] = raw_maf_data.loc[:,"n_alt_count"].round(0).astype(int)
+                                            raw_maf_data.loc[:,allelic_support_tags['control_af_tag']].astype(float) * raw_maf_data.loc[:,"n_depth"].astype(int)
                                     
+                                        raw_maf_data.loc[:,"n_alt_count"] = round(raw_maf_data.loc[:,"n_alt_count"].astype(float),0).astype(int)
+                                        raw_maf_data['n_ref_count'] = None
                                         raw_maf_data.loc[:,"n_ref_count"] = \
-                                            raw_maf_data.loc[:,"n_depth"] - raw_maf_data.loc[:,"n_alt_count"]
+                                            raw_maf_data.loc[:,"n_depth"].astype(int) - raw_maf_data.loc[:,"n_alt_count"]
         
-        raw_maf_data = raw_maf_data.fillna("")
+        #raw_maf_data = raw_maf_data.fillna("")
         with open(maf_fname, 'w') as f:
             f.write(f'{header_line}\n')
         f.close()
         raw_maf_data.to_csv(maf_fname, sep="\t", index=False, mode='a')
-        remove_file(maf_tmp_fname)                     
+        if not debug:
+            remove_file(maf_tmp_fname)
